refactor: PK 컬럼 네이밍 'id' 통일화 및 token 디렉터리 생성
* PK 컬럼의 이름을 모두 id로 통일하도록 결정하였고 이를 적용하였다.
* auth 도메인의 token 디렉터리를 생성하고 적용하였다.
