+++ spadeworker/src/main/java/com/study/spadeworker/domain/article/entity/ArticleComment.java
@@ -15,18 +15,15 @@ import java.util.Objects;
 public class ArticleComment extends BaseEntity {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "comment_id")
     private Long id;
 
     @Column(nullable = false, length = 10000)
     private String content;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "article_id")
     private Article article;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "user_id")
     private User user;
 
     private ArticleComment(String content, Article article, User user) {
