+++ spadeworker/src/main/java/com/study/spadeworker/domain/auth/jwt/service/TokenService.java
@@ -1,9 +1,9 @@
 package com.study.spadeworker.domain.auth.jwt.service;
 
-import com.study.spadeworker.domain.auth.jwt.AuthToken;
-import com.study.spadeworker.domain.auth.jwt.AuthTokenProvider;
-import com.study.spadeworker.domain.auth.jwt.UserRefreshToken;
-import com.study.spadeworker.domain.auth.jwt.UserRefreshTokenRepository;
+import com.study.spadeworker.domain.auth.jwt.token.AuthToken;
+import com.study.spadeworker.domain.auth.jwt.token.AuthTokenProvider;
+import com.study.spadeworker.domain.auth.jwt.token.UserRefreshToken;
+import com.study.spadeworker.domain.auth.jwt.token.UserRefreshTokenRepository;
 import com.study.spadeworker.domain.user.entity.constant.RoleType;
 import com.study.spadeworker.global.config.properties.AppProperties;
 import io.jsonwebtoken.Claims;
