+++ spadeworker/src/main/java/com/study/spadeworker/domain/auth/jwt/token/AuthTokenProvider.java
@@ -1,4 +1,4 @@
-package com.study.spadeworker.domain.auth.jwt;
+package com.study.spadeworker.domain.auth.jwt.token;
 
 import com.study.spadeworker.domain.auth.exception.jwt.TokenValidFailedException;
 import io.jsonwebtoken.Claims;
