+++ spadeworker/src/main/java/com/study/spadeworker/domain/auth/jwt/token/UserRefreshTokenRepository.java
@@ -1,5 +1,6 @@
-package com.study.spadeworker.domain.auth.jwt;
+package com.study.spadeworker.domain.auth.jwt.token;
 
+import com.study.spadeworker.domain.auth.jwt.token.UserRefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRefreshTokenRepository extends JpaRepository<UserRefreshToken, Long> {
