+++ spadeworker/src/main/java/com/study/spadeworker/domain/user/entity/User.java
@@ -16,7 +16,6 @@ import javax.persistence.*;
 public class User extends BaseTimeEntity {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
     private Long id;
 
     @Column(nullable = false, length = 200, unique = true)
