+++ spadeworker/src/main/java/com/study/spadeworker/global/config/security/JwtConfig.java
@@ -1,6 +1,6 @@
 package com.study.spadeworker.global.config.security;
 
-import com.study.spadeworker.domain.auth.jwt.AuthTokenProvider;
+import com.study.spadeworker.domain.auth.jwt.token.AuthTokenProvider;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
